python/sample-apps/function/requirements.txt
aiohttp
opentelemetry-sdk==1.12.0rc2
opentelemetry-api==1.12.0rc2
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc2
